Add support for a ConnectAfter function that runs an event handler after the default instead of before
